Created wrapper class which contains information to rehydrate; Moved around dependencies to create checkpointer internally
