+++ sdk/storage/Azure.Storage.DataMovement/api/Azure.Storage.DataMovement.netstandard2.0.cs
@@ -155,6 +155,14 @@ namespace Azure.Storage.DataMovement.Models
         public ReadStreamStorageResourceResult(System.IO.Stream content) { }
         public ReadStreamStorageResourceResult(System.IO.Stream content, string contentRange, string acceptRanges, byte[] rangeContentHash, Azure.Storage.DataMovement.Models.StorageResourceProperties properties) { }
     }
+    public partial class ResourceProperties
+    {
+        protected ResourceProperties() { }
+        public Azure.Storage.DataMovement.Models.TransferCheckpointerOptions Checkpointer { get { throw null; } set { } }
+        public string DestinationScheme { get { throw null; } set { } }
+        public string SourceScheme { get { throw null; } set { } }
+        public string TransferId { get { throw null; } set { } }
+    }
     public enum ServiceCopyStatus
     {
         Pending = 0,
@@ -189,6 +197,7 @@ namespace Azure.Storage.DataMovement.Models
     public partial class TransferCheckpointerOptions
     {
         public TransferCheckpointerOptions(string localCheckpointerPath) { }
+        public string CheckpointerPath { get { throw null; } }
     }
     public enum TransferCopyMethod
     {
