+++ sdk/storage/Azure.Storage.DataMovement/src/Shared/JobPlan/FolderPropertiesMode.cs
