+++ sdk/storage/Azure.Storage.DataMovement/src/Shared/JobPlan/JobPartDeleteSnapshotsOption.cs
