+++ sdk/storage/Azure.Storage.DataMovement/src/Shared/JobPlan/JobPartPermanentDeleteOption.cs
