+++ sdk/storage/Azure.Storage.DataMovement/src/Shared/JobPlan/JobPartPlanCredentialType.cs
