+++ sdk/storage/Azure.Storage.DataMovement/src/Shared/JobPlan/JobPlanOperation.cs
