+++ sdk/storage/Azure.Storage.DataMovement/src/Shared/LocalTransferCheckpointer.cs
