+++ sdk/storage/Azure.Storage.DataMovement/src/Shared/TransferCheckpointer.cs
